fix/SV-180: mock API 파일 분리
